refactor: Redis 인프라 모듈 생성 후, 로직 분리
[#64]
